Add files via upload
# ETG-Internship-Project
I had designed as a basic back-end software for use in a store like environment all we had to do was make a system that could add , remove and edit items but I was having a lot of fun so I added a shopping cart and a bunch of other features. I've also added some things like if you checkout before you quit it'll just say Thank you
but if you quit before checking out it shows the checkout.